perf: Reduce unnecessary LinkPicker re-renders
Combining state with an object in a single `useState` meant a new object
was set to the state regardless of whether new values were present. This
resulted in unexpected `act` warnings caused by unnecessary re-renders.

Refactoring LinkPicker to separate the two state values into separate
`useState` calls resolves this issue by using primitives that can be
compared directly, which results in `useState` not triggering re-renders
when the values are the same.