Add wp_print_font_faces().
This function works for both block-powered and classic sites.

For block-powered sites, it gets the fonts from theme.json via
the Font Face Resolver.

For classic sites, themes and plugins can programmatically invoke
it and pass their fonts to it.

The instance of WP_Font_Face is variable cached within this
global function instead of providing a separate global function.

Why? The entry point to use Font Face is via wp_print_font_faces().
No other methods are public. Therefore, a global wp_font_face()
function should not be needed as it would not provide benefits.